requirements.txt
faiss-cpu
huggingface_hub
ipykernel
jupyter
langchain
langchain-community
langchain-huggingface
langchain-mistralai
pandas
pypdf
python-dotenv
streamlit
sentence-transformers
sqlalchemy
tiktoken
yake
